Fix bug in Grow logic in HashTableBase
NodeLookup() returns only when a null slot is found, but if tombstones
accumulate, it can infinite loop due to no null slots existing.

This changes the hash table to track the number of tombstones, and grow
when the number of null entries is < 25%.

This bug does not affect any existing uses of PointerSet, since it
happens only for sets that alternate between add() and erase() (which
afaict, no existing uses do).

Change-Id: I846ae0468faa8084d84d0c042bfaf5a06995e4f6
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/19160
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Andrew Grieve <agrieve@google.com>